Move step code to dedicated module
This might not be necessary right now, but as more futures and state
enums are added for other operations, the `driver` module is bound to
become overloaded.
